+++ ktor-server/ktor-server-cio/common/src/io/ktor/server/cio/internal/CoroutineUtils.kt
@@ -6,7 +6,6 @@ package io.ktor.server.cio.internal
 
 import kotlinx.coroutines.*
 
-internal actual val Dispatchers.IOBridge: CoroutineDispatcher
-    get() = IO
+internal expect val Dispatchers.IOBridge: CoroutineDispatcher
 
-internal actual fun <T> runBlockingBridge(block: suspend CoroutineScope.() -> T): T = runBlocking(block = block)
+internal expect fun <T> runBlockingBridge(block: suspend CoroutineScope.() -> T): T
