--- ktor-server/ktor-server-cio/jvmAndPosix/src/io/ktor/server/cio/internal/CoroutineUtils.kt
@@ -1,9 +1,0 @@
-/*
- * Copyright 2014-2021 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
- */
-
-package io.ktor.server.cio.internal
-
-import kotlinx.coroutines.*
-
-internal expect val Dispatchers.IOBridge: CoroutineDispatcher
