+++ ktor-server/ktor-server-cio/nonJvm/src/io/ktor/server/cio/backend/SocketAddressUtils.nonJvm.kt
@@ -1,0 +1,13 @@
+/*
+ * Copyright 2014-2022 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package io.ktor.server.cio.backend
+
+import io.ktor.network.sockets.*
+import io.ktor.util.network.*
+
+internal actual fun SocketAddress.toNetworkAddress(): NetworkAddress {
+    check(this is InetSocketAddress) { "Expected inet socket address" }
+    return NetworkAddress(hostname, port)
+}
