--- ktor-server/ktor-server-test-host/posix/src/io/ktor/server/testing/Utils.posix.kt
@@ -1,8 +1,0 @@
-/*
- * Copyright 2014-2019 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
- */
-
-package io.ktor.server.testing
-
-@Suppress("ACTUAL_WITHOUT_EXPECT")
-internal actual typealias NetworkSocketTimeoutException = io.ktor.network.sockets.SocketTimeoutException
